+++ lib/olels/objects/patron_email_line.rb
@@ -33,7 +33,7 @@ module OLE_QA::Framework::OLELS
                                                             id_str = 'email_active_add_control' :
                                                             id_str = "email_active_line#{line_id - 1}_control"
                                                             b.checkbox(:id => id_str)}
-      element(:add_button)                                {b.button(:id => 'OlePatronDocument-Email_add')}
+      element(:add_button)                                {b.button(:id => 'OlePatronDocument-Email-Add_add')}
       element(:delete_button)                             {b.button(:id => "OlePatronDocument-Email_del_line#{line_id - 1}")}
     end
   end